requirements.txt
algoliasearch >= 2.0,<3.0
beautifulsoup4 == 4.7.1
cloudinary == 1.15.0
feedparser == 5.2.1
lxml == 4.3.3
pony == 0.7.10
psycopg2-binary == 2.8.2
pysocks == 1.6.8
python-telegram-bot == 11.1.0
requests == 2.21.0
selenium == 3.141.0
